requirements.txt
pyserial
flask
flask-socketio
